refactor: decouple AI service into shared utility layer
